Move studio REST apis into studio part 2
