+++ app/desktop/desktop_server.py
@@ -1,18 +1,24 @@
 import contextlib
-import os
 import threading
 import time
 
 import kiln_server.server as kiln_server
 import uvicorn
 
+from app.desktop.studio_server.prompt_api import connect_prompt_api
 from app.desktop.studio_server.provider_api import connect_provider_api
+from app.desktop.studio_server.repair_api import connect_repair_api
+from app.desktop.studio_server.settings_api import connect_settings
 from app.desktop.studio_server.webhost import connect_webhost
 
 
 def make_app():
     app = kiln_server.make_app()
     connect_provider_api(app)
+    connect_prompt_api(app)
+    connect_repair_api(app)
+    connect_settings(app)
+
     # Important: webhost must be last, it handles all other URLs
     connect_webhost(app)
     return app
