+++ app/desktop/studio_server/test_settings_api.py
@@ -6,7 +6,8 @@ import yaml
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from kiln_ai.utils.config import Config
-from kiln_server.settings import connect_settings
+
+from app.desktop.studio_server.settings_api import connect_settings
 
 
 @pytest.fixture
