+++ libs/server/kiln_server/server.py
@@ -6,10 +6,7 @@ from fastapi.middleware.cors import CORSMiddleware
 
 from .custom_errors import connect_custom_errors
 from .project_api import connect_project_api
-from .prompt_api import connect_prompt_api
-from .repair_api import connect_repair_api
 from .run_api import connect_run_api
-from .settings import connect_settings
 from .task_api import connect_task_api
 
 
@@ -36,9 +33,6 @@ def make_app():
     connect_project_api(app)
     connect_task_api(app)
     connect_run_api(app)
-    connect_repair_api(app)
-    connect_prompt_api(app)
-    connect_settings(app)
     connect_custom_errors(app)
 
     return app
